+++ docs/sphinx/conf.py
@@ -15,22 +15,15 @@ import os
 
 from pkg_resources import parse_version
 
-try:
-    from drift import __version__
-except ModuleNotFoundError:
-    from sdsstools import get_package_version
-    __version__ = get_package_version(__file__, 'sdss-drift') or 'dev'
+from drift import __version__
 
 
 # Are we building in RTD?
-on_rtd = os.environ.get('READTHEDOCS') == 'True'
+on_rtd = os.environ.get("READTHEDOCS") == "True"
 
 # Sphinx template selected in cookiecutter and whether to use releases
-sphinx_template = 'alabaster'
-use_releases = 'yes'
-
-if sphinx_template == 'sphinx-bootstrap':
-    import sphinx_bootstrap_theme
+sphinx_template = "alabaster"
+use_releases = "yes"
 
 
 # Importing matplotlib here with agg to prevent tkinter error in readthedocs
@@ -47,17 +40,23 @@ if sphinx_template == 'sphinx-bootstrap':
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ['sphinx.ext.autodoc', 'sphinx.ext.napoleon', 'sphinx.ext.autosummary',
-              'sphinx.ext.todo', 'sphinx.ext.viewcode', 'sphinx.ext.mathjax',
-              'sphinx.ext.intersphinx']
+extensions = [
+    "sphinx.ext.autodoc",
+    "sphinx.ext.napoleon",
+    "sphinx.ext.autosummary",
+    "sphinx.ext.todo",
+    "sphinx.ext.viewcode",
+    "sphinx.ext.mathjax",
+    "sphinx.ext.intersphinx",
+]
 
 # Add any paths that contain templates here, relative to this directory.
-templates_path = ['_templates']
+templates_path = ["_templates"]
 
 # The suffix(es) of source filenames.
 # You can specify multiple suffix as a list of string:
 #
-source_suffix = ['.rst', '.md']
+source_suffix = [".rst", ".md"]
 # source_suffix = '.rst'
 
 # source_parsers = {
@@ -65,19 +64,19 @@ source_suffix = ['.rst', '.md']
 # }
 
 # The master toctree document.
-master_doc = 'index'
+master_doc = "index"
 
 # General information about the project.
-project = 'drift'
-copyright = '{0}, {1}'.format('2020', 'José Sánchez-Gallego')
-author = 'José Sánchez-Gallego'
+project = "drift"
+copyright = "{0}, {1}".format("2020", "José Sánchez-Gallego")
+author = "José Sánchez-Gallego"
 
 # The version info for the project you're documenting, acts as replacement for
 # |version| and |release|, also used in various other places throughout the
 # built documents.
 
 # The short X.Y version.
-version = parse_version(__version__).base_version
+version = parse_version(__version__).base_version  # type: ignore
 # The full version, including alpha/beta/rc tags.
 release = __version__
 
@@ -91,11 +90,11 @@ language = None
 # List of patterns, relative to source directory, that match files and
 # directories to ignore when looking for source files.
 # This patterns also effect to html_static_path and html_extra_path
-exclude_patterns = ['_build', 'Thumbs.db', '.DS_Store']
+exclude_patterns = ["_build", "Thumbs.db", ".DS_Store"]
 
 # The reST default role (used for this markup: `text`) to use for all
 # documents.
-default_role = 'py:obj'
+default_role = "py:obj"
 
 # If true, '()' will be appended to :func: etc. cross-reference text.
 # add_function_parentheses = True
@@ -109,7 +108,7 @@ add_module_names = True
 # show_authors = False
 
 # The name of the Pygments (syntax highlighting) style to use.
-pygments_style = 'sphinx'
+pygments_style = "sphinx"
 
 # A list of ignored prefixes for module index sorting.
 # modindex_common_prefix = []
@@ -121,11 +120,13 @@ pygments_style = 'sphinx'
 todo_include_todos = False
 
 # Intersphinx mappings
-intersphinx_mapping = {'python': ('https://docs.python.org/3.6', None),
-                       'pymodbus': ('https://pymodbus.readthedocs.io/en/latest', None)}
+intersphinx_mapping = {
+    "python": ("https://docs.python.org/3.6", None),
+    "pymodbus": ("https://pymodbus.readthedocs.io/en/latest", None),
+}
 
-autodoc_mock_imports = ['_tkinter']
-autodoc_member_order = 'groupwise'
+autodoc_mock_imports = ["_tkinter"]
+autodoc_member_order = "groupwise"
 
 napoleon_use_rtype = False
 napoleon_use_ivar = True
@@ -139,120 +140,36 @@ rst_epilog = f"""
 
 # -- Options for HTML output ----------------------------------------------
 
-html_css_files = [
-    'pygments.css'
-]
+html_css_files = ["pygments.css"]
 
 # The theme to use for HTML and HTML Help pages.  See the documentation for
 # a list of builtin themes.
 
-if sphinx_template == 'sphinx-bootstrap':
+html_theme = "alabaster"
 
-    html_theme = 'bootstrap'
+html_theme_options = {
+    "logo": "sdssv_logo.png",
+    "github_user": "sdss",
+    "github_repo": project,
+    "github_button": True,
+    "github_type": "star",
+    "sidebar_collapse": False,
+    "sidebar_includehidden": False,
+    "page_width": "80%",
+}
 
-    html_sidebars = {}
+html_sidebars = {
+    "**": [
+        "about.html",
+        "localtoc.html",
+        "relations.html",
+        "searchbox.html",
+    ]
+}
 
-    # Theme options are theme-specific and customize the look and feel of a theme
-    # further.  For a list of options available for each theme, see the
-    # documentation.
-    html_theme_options = {
-        # Navigation bar title. (Default: ``project`` value)
-        'navbar_title': "SDSS: {0}".format(project),
+html_css_files += ["custom.css"]
 
-        # Tab name for entire site. (Default: "Site")
-        'navbar_site_name': "Site",
-
-        # A list of tuples containing pages or urls to link to.
-        # Valid tuples should be in the following forms:
-        #    (name, page)                 # a link to a page
-        #    (name, "/aa/bb", 1)          # a link to an arbitrary relative url
-        #    (name, "http://example.com", True) # arbitrary absolute url
-        # Note the "1" or "True" value above as the third argument to indicate
-        # an arbitrary url.
-        'navbar_links': [
-        ],
-
-        # Render the next and previous page links in navbar. (Default: true)
-        'navbar_sidebarrel': False,
-
-        # Render the current pages TOC in the navbar. (Default: true)
-        'navbar_pagenav': True,
-
-        # Tab name for the current pages TOC. (Default: "Page")
-        'navbar_pagenav_name': "Page",
-
-        # Global TOC depth for "site" navbar tab. (Default: 1)
-        # Switching to -1 shows all levels.
-        'globaltoc_depth': 2,
-
-        # Include hidden TOCs in Site navbar?
-        #
-        # Note: If this is "false", you cannot have mixed ``:hidden:`` and
-        # non-hidden ``toctree`` directives in the same page, or else the build
-        # will break.
-        #
-        # Values: "true" (default) or "false"
-        'globaltoc_includehidden': "true",
-
-        # HTML navbar class (Default: "navbar") to attach to <div> element.
-        # For black navbar, do "navbar navbar-inverse"
-        'navbar_class': "navbar",
-
-        # Fix navigation bar to top of page?
-        # Values: "true" (default) or "false"
-        'navbar_fixed_top': "true",
-
-        # Location of link to source.
-        # Options are "nav" (default), "footer" or anything else to exclude.
-        'source_link_position': "",
-
-        # Bootswatch (http://bootswatch.com/) theme.
-        #
-        # Options are nothing (default) or the name of a valid theme
-        # such as "amelia" or "cosmo".
-        'bootswatch_theme': "paper",
-
-        # Choose Bootstrap version.
-        # Values: "3" (default) or "2" (in quotes)
-        'bootstrap_version': "3",
-    }
-
-    # Add any paths that contain custom themes here, relative to this directory.
-    html_theme_path = sphinx_bootstrap_theme.get_html_theme_path()
-
-    html_logo = '_static/sdssv_logo_small.png'
-
-    html_css_files += ["custom_bootstrap.css"]
-
-    html_sidebars = {'**': ['localtoc.html']}
-
-elif sphinx_template == 'alabaster':
-
-    html_theme = 'alabaster'
-
-    html_theme_options = {
-        'logo': 'sdssv_logo.png',
-        'github_user': 'sdss',
-        'github_repo': project,
-        'github_button': True,
-        'github_type': 'star',
-        'sidebar_collapse': False,
-        'sidebar_includehidden': False,
-        'page_width': '80%'
-    }
-
-    html_sidebars = {
-        '**': [
-            'about.html',
-            'localtoc.html',
-            'relations.html',
-            'searchbox.html',
-        ]
-    }
-
-    html_css_files += ["custom.css"]
-
-html_favicon = './_static/favicon_sdssv.ico'
+html_favicon = "./_static/favicon_sdssv.ico"
 
 # Add any paths that contain custom static files (such as style sheets) here,
 # relative to this directory. They are copied after the builtin static files,
@@ -262,13 +179,13 @@ html_favicon = './_static/favicon_sdssv.ico'
 if on_rtd:
     html_static_path = []
 else:
-    html_static_path = ['_static']
+    html_static_path = ["_static"]
 
 
 # -- Options for HTMLHelp output ------------------------------------------
 
 # Output file base name for HTML help builder.
-htmlhelp_basename = '{0}pdoc'.format('drift')
+htmlhelp_basename = "{0}pdoc".format("drift")
 
 
 # -- Options for LaTeX output ---------------------------------------------
@@ -277,15 +194,12 @@ latex_elements = {
     # The paper size ('letterpaper' or 'a4paper').
     #
     # 'papersize': 'letterpaper',
-
     # The font size ('10pt', '11pt' or '12pt').
     #
     # 'pointsize': '10pt',
-
     # Additional stuff for the LaTeX preamble.
     #
     # 'preamble': '',
-
     # Latex figure (float) alignment
     #
     # 'figure_align': 'htbp',
@@ -295,18 +209,20 @@ latex_elements = {
 # (source start file, target name, title,
 #  author, documentclass [howto, manual, or own class]).
 latex_documents = [
-    (master_doc, '{0}.tex'.format(project), u'{0} Documentation'.format(project),
-     author, 'manual'),
+    (
+        master_doc,
+        "{0}.tex".format(project),
+        "{0} Documentation".format(project),
+        author,
+        "manual",
+    ),
 ]
 
 # -- Options for manual page output ---------------------------------------
 
 # One entry per manual page. List of tuples
 # (source start file, name, description, authors, manual section).
-man_pages = [
-    (master_doc, 'drift', u'{0} Documentation'.format(project),
-     [author], 1)
-]
+man_pages = [(master_doc, "drift", "{0} Documentation".format(project), [author], 1)]
 
 # -- Options for Texinfo output -------------------------------------------
 
@@ -314,15 +230,21 @@ man_pages = [
 # (source start file, target name, title, author,
 #  dir menu entry, description, category)
 texinfo_documents = [
-    (master_doc, project, u'{0} Documentation'.format(project),
-     author, project, 'One line description of project.',
-     'Miscellaneous'),
+    (
+        master_doc,
+        project,
+        "{0} Documentation".format(project),
+        author,
+        project,
+        "One line description of project.",
+        "Miscellaneous",
+    ),
 ]
 
-if use_releases == 'yes':
+if use_releases == "yes":
 
-    extensions += ['sdsstools.releases']
+    extensions += ["sdsstools.releases"]
 
-    releases_github_path = 'sdss/drift'
-    releases_document_name = ['CHANGELOG']
+    releases_github_path = "sdss/drift"
+    releases_document_name = ["CHANGELOG"]
     releases_unstable_prehistory = True