+++ python/drift/drift.py
@@ -6,34 +6,29 @@
 # @Filename: drift.py
 # @License: BSD 3-clause (http://www.opensource.org/licenses/BSD-3-Clause)
 
+from __future__ import annotations
+
 import asyncio
 import importlib
 import warnings
 from collections import defaultdict
 
+from typing import Callable, Optional, Type, cast
+
 from pymodbus.client.asynchronous.async_io import AsyncioModbusTcpClient
 from yaml import SafeLoader, load
 
-from . import adaptors, log
-from .exceptions import DriftError, DriftUserWarning
+from drift import adaptors, log
+from drift.exceptions import DriftError, DriftUserWarning
 
 
-__ALL__ = ['Device', 'Relay', 'Module', 'MODULES', 'Drift']
+__all__ = ["Device", "Relay", "Module", "MODULES", "Drift"]
 
 
 MODULES = {
-    '750-450': {
-        'mode': 'input',
-        'channels': 4
-    },
-    '750-497': {
-        'mode': 'input',
-        'channels': 8
-    },
-    '750-530': {
-        'mode': 'output',
-        'channels': 8
-    }
+    "750-450": {"mode": "input", "channels": 4},
+    "750-497": {"mode": "input", "channels": 8},
+    "750-530": {"mode": "output", "channels": 8},
 }
 
 
@@ -84,28 +79,36 @@ class Module(object):
 
     Parameters
     ----------
-    name : str
+    name
         A name to be associated with this module.
-    drift : .Drift
+    drift
         The instance of `.Drift` used to communicate with the modbus network.
-    address : int
+    address
         The initial address associated with this module. In modbus, addresses
         start at 40001.
-    model : str
+    model
         The Drift model for this module.
-    mode : str
+    mode
         Whether this is an ``input`` or ``output`` module. If not provided it
         will be determined from ``model``.
-    channels : int
+    channels
         The number of channels in this module. If not provided it will be
         determined from ``model``.
-    description : str
+    description
         A description or comment for this module.
 
     """
 
-    def __init__(self, name, drift, address, model=None, mode=None,
-                 channels=None, description=''):
+    def __init__(
+        self,
+        name: str,
+        drift: Drift,
+        address: int,
+        model: Optional[str] = None,
+        mode: Optional[str] = None,
+        channels: Optional[int] = None,
+        description: str = "",
+    ):
 
         self.name = name
         self.drift = drift
@@ -117,10 +120,10 @@ class Module(object):
 
         if self.model:
             if self.model in MODULES:
-                default_mode = MODULES[self.model]['mode']
-                default_channels = MODULES[self.model]['channels']
+                default_mode = MODULES[self.model]["mode"]
+                default_channels = MODULES[self.model]["channels"]
             else:
-                raise DriftError(f'Unknown model {self.model!r}')
+                raise DriftError(f"Unknown model {self.model!r}")
         else:
             default_mode = None
             default_channels = None
@@ -128,57 +131,74 @@ class Module(object):
         if mode is not None:
             self.mode = mode
             if default_mode and default_mode != self.mode:
-                warnings.warn(f'mode {self.mode!r} is different from '
-                              f'default mode {default_mode!r} for model '
-                              f'{self.model},', DriftUserWarning)
+                warnings.warn(
+                    f"mode {self.mode!r} is different from "
+                    f"default mode {default_mode!r} for model "
+                    f"{self.model},",
+                    DriftUserWarning,
+                )
         else:
             if default_mode is None:
-                raise DriftError('Cannot determine module mode.')
+                raise DriftError("Cannot determine module mode.")
             self.mode = default_mode
 
-        if self.mode not in ['input', 'output']:
-            raise DriftError(f'Invalid mode {mode}.')
+        if self.mode not in ["input", "output"]:
+            raise DriftError(f"Invalid mode {mode}.")
 
         if channels is not None:
             self.channels = channels
             if default_channels and default_channels != self.channels:
-                warnings.warn(f'channels {self.channels!r} is different from '
-                              f'default channels {default_channels!r} for '
-                              f'model {self.model},', DriftUserWarning)
+                warnings.warn(
+                    f"channels {self.channels!r} is different from "
+                    f"default channels {default_channels!r} for "
+                    f"model {self.model},",
+                    DriftUserWarning,
+                )
         else:
             if default_channels is None:
-                raise DriftError('Cannot determine module channels.')
+                raise DriftError("Cannot determine module channels.")
             self.channels = default_channels
 
         self.devices = CaseInsensitiveDict()
 
-        log.info(f'Created module {self.name}@{self.address} with '
-                 f'mode {self.mode} and {self.channels} channels.')
+        log.info(
+            f"Created module {self.name}@{self.address} with "
+            f"mode {self.mode} and {self.channels} channels."
+        )
 
     def __repr__(self):
-        return (f'<Module {self.name}@{self.address} (mode={self.mode!r}, '
-                f'channels={self.channels}, n_devices={len(self.devices)})>')
+        return (
+            f"<Module {self.name}@{self.address} (mode={self.mode!r}, "
+            f"channels={self.channels}, n_devices={len(self.devices)})>"
+        )
 
-    def __getitem__(self, name):
+    def __getitem__(self, name: str) -> Device:
         """Gets a device."""
 
         return self.devices[name]
 
-    def add_device(self, name, channel=None, device_class=None, **kwargs):
+    def add_device(
+        self,
+        name: str | Device,
+        channel: Optional[int] = None,
+        device_class: Optional[Type[Device]] = None,
+        **kwargs,
+    ) -> Device:
         """Adds a device
 
         Parameters
         ----------
-        name : str or .Device
+        name
             The name of the device. It is treated as case-insensitive. It can
             also be a `.Device` instance, in which case the remaining
             parameters will be ignored.
-        channel : int
+        channel
             The channel of the device in the module (relative to the
-            module address), zero-indexed.
-        device_class : class
+            module address), zero-indexed. Required if ``name`` is not
+            a `.Device`.
+        device_class
             Either `.Device` or a subclass of it.
-        kwargs : dict
+        kwargs
             Other parameters to pass to `.Device`.
 
         """
@@ -192,28 +212,31 @@ class Module(object):
             device_class = device_class or Device
 
             if name in self.devices:
-                raise DriftError(f'Device {name!r} already exists in '
-                                 f'module {self.name!r}.')
+                raise DriftError(
+                    f"Device {name!r} already exists in " f"module {self.name!r}."
+                )
+
+            assert channel is not None, "channel is not defined for the new device."
 
             device = device_class(self, name, channel, **kwargs)
 
         self.devices[device.name] = device
-        log.info(f'Added device {device.name} to module {self.name}.')
+        log.info(f"Added device {device.name} to module {self.name}.")
 
         return device
 
-    def remove_device(self, name):
+    def remove_device(self, name: str) -> Device:
         """Removes a device.
 
         Parameters
         ----------
-        name : str
+        name
             The name of the device to remove.
 
         """
 
         device = self.devices.pop(name)
-        log.info(f'Removed device {name} from module {self.name}.')
+        log.info(f"Removed device {name} from module {self.name}.")
 
         return device
 
@@ -223,25 +246,25 @@ class Device(object):
 
     Parameters
     ----------
-    module : .Module
+    module
         The `.Module` to which this device is connected to.
-    name : str
+    name
         The name of the device. It is treated as case-insensitive.
-    channel : int
+    channel
         The channel of the PLC inside the module, zero-indexed, so that the
         first channel is 0 and the full address of the device is
         ``module_address + channel - 40001``.
-    coils : bool
+    coils
         Whether to read coils or registers for this device.
-    category : str
+    category
         A user-defined category for this device. This can be used to group
         different devices of similar type, for example ``'temperature'`` to
         indicate temperature sensors.
-    description : str
+    description
         A description of the purpose of the device.
-    units : str
+    units
         The units of the values returned, if an adaptor is used.
-    adaptor : str, dict, or function
+    adaptor
         The adaptor to be used to convert read registers to physical values.
         It can be a string indicating one of the provided :ref:`adaptors
         <adaptors>`, a string with the format ``module1.module2:function``
@@ -255,10 +278,19 @@ class Device(object):
 
     __type__ = None
 
-    def __init__(self, module, name, channel, coils=False, description='',
-                 units=None, category=None, adaptor=None):
+    def __init__(
+        self,
+        module: Module,
+        name: str,
+        channel: int,
+        coils: bool = False,
+        description: str = "",
+        units: Optional[str] = None,
+        category: Optional[str] = None,
+        adaptor: Optional[str | dict | Callable] = None,
+    ):
 
-        assert isinstance(module, Module), 'module not a valid Module object.'
+        assert isinstance(module, Module), "module not a valid Module object."
 
         self.module = module
         self.name = name
@@ -269,13 +301,13 @@ class Device(object):
         self.category = category
         self.adaptor = self._parse_adaptor(adaptor)
 
-        log.info(f'Created device '
-                 f'{self.name}@{self.module.address}:{self.channel}.')
+        log.info(f"Created device {self.name}@{self.module.address}:{self.channel}.")
 
     def __repr__(self):
-        return (f'<Device {self.name}@{self.module.address}:{self.channel}>')
+        return f"<Device {self.name}@{self.module.address}:{self.channel}>"
 
-    def _parse_adaptor(self, adaptor):
+    @staticmethod
+    def _parse_adaptor(adaptor):
         """Parses the adaptor."""
 
         if adaptor is None:
@@ -283,17 +315,17 @@ class Device(object):
 
         elif isinstance(adaptor, str):
 
-            if ':' in adaptor:
+            if ":" in adaptor:
                 try:
-                    module_str, adaptor = adaptor.split(':')
+                    module_str, adaptor = adaptor.split(":")
                 except ValueError:
-                    raise DriftError(f'Badly formatted adaptor {adaptor}.')
+                    raise DriftError(f"Badly formatted adaptor {adaptor}.")
                 module = importlib.import_module(module_str)
             else:
                 module = adaptors
 
             if hasattr(module, adaptor) is False:
-                raise DriftError(f'Cannot find adaptor {adaptor}.')
+                raise DriftError(f"Cannot find adaptor {adaptor}.")
             else:
                 return getattr(module, adaptor)
 
@@ -301,18 +333,18 @@ class Device(object):
             return dict(adaptor)
 
     @property
-    def address(self):
+    def address(self) -> int:
         """Returns the full address of this PLC."""
 
         return self.module.address + self.channel - 40001
 
     @property
-    def client(self):
+    def client(self) -> AsyncioModbusTcpClient | None:
         """Returns the ``pymodbus`` client."""
 
         return self.module.drift.client
 
-    async def read(self, adapt=True):
+    async def read(self, adapt=True) -> tuple[int | float, None | str]:
         """Reads the value of the coil or register.
 
         If ``adapt=True`` and a valid adaptor was provided, the value returned
@@ -340,8 +372,10 @@ class Device(object):
             resp = await reader(self.address, count=1)
 
             if resp.function_code > 0x80:
-                raise DriftError(f'Invalid response for device '
-                                 f'{self.name!r}: 0x{resp.function_code:02X}.')
+                raise DriftError(
+                    f"Invalid response for device "
+                    f"{self.name!r}: 0x{resp.function_code:02X}."
+                )
 
             value = resp.registers[0] if not self.coils else resp.bits[0]
 
@@ -353,8 +387,10 @@ class Device(object):
                     value = self.adaptor(value)
                 else:
                     if value not in self.adaptor:
-                        raise DriftError(f'Cannot find associated value for '
-                                         f'{value} in adaptor mapping.')
+                        raise DriftError(
+                            f"Cannot find associated value for "
+                            f"{value} in adaptor mapping."
+                        )
                     else:
                         value = self.adaptor[value]
 
@@ -365,11 +401,11 @@ class Device(object):
 
         return value, units
 
-    async def write(self, value):
+    async def write(self, value) -> bool:
         """Writes values to a coil or register."""
 
-        if self.module.mode != 'output':
-            raise DriftError('Writing is not allowed to this module.')
+        if self.module.mode != "output":
+            raise DriftError("Writing is not allowed to this module.")
 
         async with self.module.drift:
 
@@ -381,8 +417,10 @@ class Device(object):
                 resp = await protocol.write_register(self.address, value)
 
             if resp.function_code > 0x80:
-                raise DriftError(f'Invalid response for device {self.name!r}: '
-                                 f'0x{resp.function_code:02X}.')
+                raise DriftError(
+                    f"Invalid response for device {self.name!r}: "
+                    f"0x{resp.function_code:02X}."
+                )
 
         return True
 
@@ -396,28 +434,28 @@ class Relay(Device):
 
     """
 
-    __type__ = 'relay'
+    __type__ = "relay"
 
-    def __init__(self, *args, relay_type='NC', **kwargs):
+    def __init__(self, *args, relay_type="NC", **kwargs):
 
         self.relay_type = relay_type
 
-        if relay_type == 'NC':
-            adaptor = [(False, 'closed'), (True, 'open')]
+        if relay_type == "NC":
+            adaptor = [(False, "closed"), (True, "open")]
         else:
-            adaptor = [(False, 'open'), (True, 'closed')]
+            adaptor = [(False, "open"), (True, "closed")]
 
-        kwargs['adaptor'] = adaptor
+        kwargs["adaptor"] = adaptor
 
-        if 'coils' not in kwargs:
-            kwargs['coils'] = True
+        if "coils" not in kwargs:
+            kwargs["coils"] = True
 
         super().__init__(*args, **kwargs)
 
     async def open(self):
         """Opens the relay."""
 
-        if self.relay_type == 'NC':
+        if self.relay_type == "NC":
             await self.write(True)
         else:
             await self.write(False)
@@ -425,7 +463,7 @@ class Relay(Device):
     async def close(self):
         """Closes the relay."""
 
-        if self.relay_type == 'NC':
+        if self.relay_type == "NC":
             await self.write(False)
         else:
             await self.write(True)
@@ -452,9 +490,6 @@ class Drift(object):
         The IP of the TCP modbus server.
     port : int
         The port of the TCP modbus server.
-    loop
-        The event loop to use.
-
 
     The `.Drift` manages the TCP connection to the modbus ethernet module
     using `Pymodbus <pymodbus.readthedocs.io/en/latest/index.html>`__. The
@@ -473,17 +508,16 @@ class Drift(object):
 
     """
 
-    def __init__(self, address, port=502, loop=None):
+    def __init__(self, address: str, port: int = 502):
 
         self.address = address
         self.port = port
-        self.client = AsyncioModbusTcpClient(address, port=port, loop=loop)
-        self.loop = self.client.loop
+        self.client = AsyncioModbusTcpClient(address, port=port)
 
         self.modules = CaseInsensitiveDict()
 
     def __repr__(self):
-        return f'<Drift @ {self.address}>'
+        return f"<Drift @ {self.address}>"
 
     async def __aenter__(self):
         """Initialises the connection to the server."""
@@ -491,12 +525,12 @@ class Drift(object):
         try:
             await asyncio.wait_for(self.client.connect(), timeout=1)
         except asyncio.TimeoutError:
-            raise DriftError(f'Failed connecting to server at {self.address}.')
+            raise DriftError(f"Failed connecting to server at {self.address}.")
 
         if not self.client.connected:
-            raise DriftError(f'Failed connecting to server at {self.address}.')
+            raise DriftError(f"Failed connecting to server at {self.address}.")
 
-        log.debug(f'Connected to {self.address}.')
+        log.debug(f"Connected to {self.address}.")
 
         return
 
@@ -505,38 +539,38 @@ class Drift(object):
 
         self.client.stop()
 
-        log.debug(f'Disonnected from {self.address}.')
+        log.debug(f"Disonnected from {self.address}.")
 
         return
 
-    def __getitem__(self, name):
+    def __getitem__(self, name: str):
         """Gets a module."""
 
         return self.modules[name]
 
-    def add_module(self, name, address, **params):
+    def add_module(self, name: str, address: int, **params) -> Module:
         """Adds a new module.
 
         Parameters
         ----------
-        name : str
+        name
             The name of the module.
-        address : int
+        address
             The modbus address.
-        params : dict
+        params
             Arguments to be passed to `.Module` for initialisation.
 
         """
 
         if name in self.modules:
-            raise ValueError(f'Module {name} is already connected.')
+            raise ValueError(f"Module {name} is already connected.")
 
         self.modules[name] = Module(name, self, address, **params)
 
         return self.modules[name]
 
-    def get_device(self, device):
-        """Gets the `.Module` instance that matches ``device``.
+    def get_device(self, device: str) -> Device:
+        """Gets the `.Device` instance that matches ``device``.
 
         If the case-insensitive name of the device is unique within the pool
         of connected devices, the device name is sufficient. Otherwise the
@@ -545,8 +579,8 @@ class Drift(object):
 
         """
 
-        if '.' in device:
-            module, device = device.split('.')
+        if "." in device:
+            module, device = device.split(".")
             return self.modules[module].devices[device]
 
         device = device.lower()
@@ -557,26 +591,28 @@ class Drift(object):
                 dev_to_mod[dev.lower()].append(module.lower())
 
         if device not in dev_to_mod:
-            raise ValueError(f'Device {device} is not connected.')
+            raise ValueError(f"Device {device} is not connected.")
 
         elif len(dev_to_mod[device]) == 1:
             return self.modules[dev_to_mod[device.lower()][0]].devices[device]
 
         else:
-            raise ValueError(f'Multiple devices with name {device} found. '
-                             'Use a module-qualified name.')
+            raise ValueError(
+                f"Multiple devices with name {device} found. "
+                "Use a module-qualified name."
+            )
 
-    async def read_device(self, name, adapt=True):
+    async def read_device(self, name: str, adapt: bool = True):
         """Reads a device.
 
         Parameters
         ----------
-        adapt : bool
+        adapt
             If possible, convert the value to real units.
 
         Returns
         -------
-        read_value : tuple
+        read_value
             The read value and its associated units.
 
         """
@@ -610,14 +646,14 @@ class Drift(object):
 
         for module in self.modules:
             for device in self.modules[module].devices.values():
-                name = f'{module}.{device.name.lower()}'
+                name = f"{module}.{device.name.lower()}"
                 if device.category and device.category.lower() == category:
                     values[name] = await device.read(adapt=adapt)
 
         return values
 
     @classmethod
-    def from_config(cls, config):
+    def from_config(cls, config: str | dict):
         """Loads a Drift from a dictionary or YAML file.
 
         Parameters
@@ -630,26 +666,26 @@ class Drift(object):
         """
 
         if isinstance(config, str):
-            config = load(open(config, 'r'), Loader=SafeLoader)
+            config = cast(dict, load(open(config, "r"), Loader=SafeLoader))
 
-        address = config['address']
-        port = config.get('port', 502)
+        address = config["address"]
+        port = config.get("port", 502)
 
         new_drift = cls(address, port)
 
-        for module in config.get('modules', {}):
+        for module in config.get("modules", {}):
 
-            module_config = config['modules'][module]
-            devices = module_config.pop('devices', {})
+            module_config = config["modules"][module]
+            devices = module_config.pop("devices", {})
 
             new_drift.add_module(module, **module_config)
 
             for device in devices:
 
                 device_config = devices[device].copy()
-                channel = device_config.pop('channel')
+                channel = device_config.pop("channel")
                 device_class = None
-                type_ = device_config.pop('type', None)
+                type_ = device_config.pop("type", None)
 
                 if type_:
                     for subclass in Device.__subclasses__():
@@ -660,11 +696,12 @@ class Drift(object):
                     device_class = Device
 
                 if device_class is None:
-                    raise DriftError('Cannot find valid device class for '
-                                     f'type {type_}.')
+                    raise DriftError(
+                        "Cannot find valid device class for " f"type {type_}."
+                    )
 
-                new_drift.modules[module].add_device(device, channel,
-                                                     device_class,
-                                                     **device_config)
+                new_drift.modules[module].add_device(
+                    device, channel, device_class, **device_config
+                )
 
         return new_drift