+++ CHANGELOG.rst
@@ -4,6 +4,8 @@
 Change Log
 ==========
 
+* :support:`4` Add typing and use Black formatting.
+
 * :release:`0.1.5 <2020-10-10>`
 * :bug:`3` Fix import of ``AsyncioModbusTcpClient`` when using ``pymodbus>=2.4.0``.